# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BDE Farm Trac API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /roles:
    get:
      operationId: listRoles
      tags: [roles]
      summary: List roles (system + tenant-scoped)
      parameters:
        - $ref: "#/components/parameters/TenantSlugHeader"
      responses:
        "200":
          description: Role list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoleListResponse"
    post:
      operationId: createRole
      tags: [roles]
      summary: Create a custom role for the tenant
      parameters:
        - $ref: "#/components/parameters/TenantSlugHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRoleBody"
      responses:
        "201":
          description: Role created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoleEnvelope"
...
  /roles/{roleId}/permissions:
    get:
      operationId: getRolePermissions
      tags: [roles]
      summary: Get permissions for a role
      parameters:
        - $ref: "#/components/parameters/TenantSlugHeader"
        - name: roleId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Permission list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PermissionListResponse"
  /roles/{roleId}/permissions/{moduleId}:
    put:
      operationId: updateRolePermission
      tags: [roles]
      summary: Update permission for a role on a module
      parameters:
        - $ref: "#/components/parameters/TenantSlugHeader"
        - name: roleId
          in: path
          required: true
          schema:
            type: integer
        - name: moduleId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePermissionBody"
      responses:
        "200":
          description: Permission updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PermissionEnvelope"
        "201":
          description: Permission created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PermissionEnvelope"
  /billing/checkout:
    post:
      operationId: createCheckoutSession
      tags: [billing]
      summary: Create a Stripe checkout session for module subscriptions
      parameters:
        - $ref: "#/components/parameters/TenantSlugHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutBody"
      responses:
        "200":
          description: Checkout URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutResponse"
  /billing/subscriptions:
    get:
      operationId: listSubscriptions
      tags: [billing]
      summary: List subscriptions for the current tenant
      parameters:
        - $ref: "#/components/parameters/TenantSlugHeader"
      responses:
        "200":
          description: Subscription list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionListResponse"
  /billing/webhook:
    post:
      operationId: stripeWebhook
      tags: [billing]
      summary: Stripe webhook endpoint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Webhook received
          content:
            application/json:
              schema:
                type: object
                properties:
                  received:
                    type: boolean
                required:
                  - received
...
  /admin/tenants/{tenantId}:
    get:
      operationId: adminGetTenantDetail
      tags: [admin]
      summary: Get tenant detail with farms, subscriptions, users
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tenant detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminTenantDetailResponse"
...
  /admin/impersonate:
    post:
      operationId: adminImpersonate
      tags: [admin]
      summary: Begin impersonating a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImpersonateBody"
      responses:
        "200":
          description: Impersonation context
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImpersonateResponse"
components:
  parameters:
...
    TenantSlugHeader:
      name: x-tenant-slug
      in: header
      required: true
      description: Tenant slug to scope the request
      schema:
        type: string
  schemas:
...
    Tenant:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        isActive:
          type: boolean
        stripeCustomerId:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, name, slug, contactEmail, isActive, createdAt]
...
    Farm:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        name:
          type: string
        address:
          type: ["string", "null"]
        postcode:
          type: ["string", "null"]
        cphNumber:
          type: ["string", "null"]
        gridReference:
          type: ["string", "null"]
        totalAcreage:
          type: ["integer", "null"]
        sectorArable:
          type: boolean
        sectorBeef:
          type: boolean
        sectorDairy:
          type: boolean
        sectorPigs:
          type: boolean
        sectorPoultry:
          type: boolean
        sectorHorticulture:
          type: boolean
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, tenantId, name, isActive]
...
    Role:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: ["integer", "null"]
        name:
          type: string
        description:
          type: ["string", "null"]
        isSystemRole:
          type: boolean
      required: [id, name, isSystemRole]
    RoleListResponse:
      type: object
      properties:
        roles:
          type: array
          items:
            $ref: "#/components/schemas/Role"
      required: [roles]
    RoleEnvelope:
      type: object
      properties:
        role:
          $ref: "#/components/schemas/Role"
      required: [role]
    CreateRoleBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
      required: [name]
...
    Permission:
      type: object
      properties:
        id:
          type: integer
        roleId:
          type: integer
        moduleId:
          type: integer
        farmId:
          type: ["integer", "null"]
        moduleName:
          type: string
        canRead:
          type: boolean
        canWrite:
          type: boolean
        canDelete:
          type: boolean
        canApprove:
          type: boolean
      required: [id, roleId, moduleId, canRead, canWrite, canDelete, canApprove]
    PermissionListResponse:
      type: object
      properties:
        permissions:
          type: array
          items:
            $ref: "#/components/schemas/Permission"
      required: [permissions]
    PermissionEnvelope:
      type: object
      properties:
        permission:
          $ref: "#/components/schemas/Permission"
      required: [permission]
    UpdatePermissionBody:
      type: object
      properties:
        farmId:
          type: ["integer", "null"]
          description: Optional farm ID for farm-scoped permissions. Null for tenant-wide.
        canRead:
          type: boolean
        canWrite:
          type: boolean
        canDelete:
          type: boolean
        canApprove:
          type: boolean
    CheckoutBody:
      type: object
      properties:
        farmId:
          type: integer
        moduleIds:
          type: array
          items:
            type: integer
      required: [farmId, moduleIds]
    CheckoutResponse:
      type: object
      properties:
        checkoutUrl:
          type: string
      required: [checkoutUrl]
    Subscription:
      type: object
      properties:
        id:
          type: integer
        farmId:
          type: integer
        moduleId:
          type: integer
        moduleName:
          type: string
        status:
          type: string
        currentPeriodEnd:
          type: ["string", "null"]
          format: date-time
      required: [id, farmId, moduleId, moduleName, status]
    SubscriptionListResponse:
      type: object
      properties:
        subscriptions:
          type: array
          items:
            $ref: "#/components/schemas/Subscription"
      required: [subscriptions]
...
    AdminTenantDetailResponse:
      type: object
      properties:
        tenant:
          $ref: "#/components/schemas/Tenant"
        farms:
          type: array
          items:
            $ref: "#/components/schemas/Farm"
        subscriptions:
          type: array
          items:
            $ref: "#/components/schemas/Subscription"
        users:
          type: array
          items:
            type: object
            properties:
              userId:
                type: string
              roleId:
                type: integer
              isActive:
                type: boolean
              email:
                type: ["string", "null"]
              firstName:
                type: ["string", "null"]
              lastName:
                type: ["string", "null"]
            required: [userId, roleId, isActive]
      required: [tenant, farms, subscriptions, users]
...
    ImpersonateBody:
      type: object
      properties:
        userId:
          type: string
        tenantId:
          type: integer
      required: [userId, tenantId]
    ImpersonateResponse:
      type: object
      properties:
        impersonation:
          type: object
          properties:
            userId:
              type: string
            email:
              type: ["string", "null"]
            firstName:
              type: ["string", "null"]
            lastName:
              type: ["string", "null"]
            tenantId:
              type: integer
            note:
              type: string
          required: [userId, tenantId, note]
      required: [impersonation]