# lib/api-zod/src/generated/types/updatePermissionBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BDE Farm Trac API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdatePermissionBody {
  /**
   * Optional farm ID for farm-scoped permissions. Null for tenant-wide.
   * @nullable
   */
  farmId?: number | null;
  canRead?: boolean;
  canWrite?: boolean;
  canDelete?: boolean;
  canApprove?: boolean;
}